+++ actito-geo/src/main/java/com/actito/geo/ActitoGeo.kt
@@ -56,8 +56,6 @@ import com.actito.geo.models.ActitoBeaconSession
 import com.actito.geo.models.ActitoLocation
 import com.actito.geo.models.ActitoRegion
 import com.actito.internal.network.request.ActitoRequest
-import com.actito.ktx.device
-import com.actito.ktx.events
 import com.actito.models.ActitoApplication
 import com.actito.utilities.collections.takeEvenlySpaced
 import com.actito.utilities.coroutines.actitoCoroutineScope
@@ -830,14 +828,14 @@ public object ActitoGeo {
                 context,
                 0,
                 locationIntent,
-                PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
+                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
             )
         } else {
             PendingIntent.getBroadcast(
                 context,
                 0,
                 locationIntent,
-                PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_UPDATE_CURRENT,
+                PendingIntent.FLAG_UPDATE_CURRENT,
             )
         }
 
@@ -853,14 +851,14 @@ public object ActitoGeo {
                 context,
                 0,
                 geofencingIntent,
-                PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
+                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
             )
         } else {
             PendingIntent.getBroadcast(
                 context,
                 0,
                 geofencingIntent,
-                PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_UPDATE_CURRENT,
+                PendingIntent.FLAG_UPDATE_CURRENT,
             )
         }