+++ actito-geo/src/main/java/com/actito/geo/internal/LaunchComponent.kt
@@ -2,11 +2,13 @@ package com.actito.geo.internal
 
 import android.content.SharedPreferences
 import android.location.Geocoder
+import androidx.annotation.Keep
 import com.actito.Actito
 import com.actito.geo.ActitoGeo
 import com.actito.geo.internal.storage.LocalStorage
 import com.actito.internal.ActitoLaunchComponent
 
+@Keep
 public class LaunchComponent : ActitoLaunchComponent {
     override fun migrate(savedState: SharedPreferences, settings: SharedPreferences) {
         val localStorage = LocalStorage(Actito.requireContext())